Rename components directory to elements
