+++ src/ui/elements/Header/PlayingBar.tsx
