+++ src/ui/elements/Header/PlayingBarInfos.tsx
