+++ src/ui/elements/Settings/CheckboxSetting.tsx
