+++ src/ui/elements/Settings/SettingsAbout.tsx
