+++ src/ui/elements/Shared/FullViewMessage.tsx
