+++ src/ui/elements/Shared/PlayingIndicator.tsx
